Create __init__.py
Custom components require this file. Copied from original goldair_heater.py file. Not sure which of the setup code is meant to be in this file, and which is meant to remain the the goldair_heater.py file, but the component works (after making some other changes in other files) using this.