Be less eager about changing protocol.
Once the device is communicating, only switch protocols after the connection retry limit is reached.
This prevents compounding the problem, in particular when v3.3 protocol devices send back responses that pytuya does not handle, as switching to v3.1 seems to cause these devices to go offline for a few minutes sometimes.